chore(apps): move the access fixture into tests, drop deleted docs paths from comments
- `src/testing/app-access-fixture.ts` was a test-only helper living in the
  shipped source tree. `src/testing/index.ts` never exported it and its only
  five importers are tests, so it moves to `tests/app-access-fixture.ts`;
  tsconfig.test.json already covers `tests`, and the build tsconfig (`src` only)
  now has one less file to compile into dist.
- `claude-turn.ts` no longer exports `ClaudeSessionInput` or `SdkModule` —
  declaration emit inlines both into `dist/claude-turn.d.ts`, so the public
  `createClaudeSession` signature is unchanged. `ClaudeSession` stays exported:
  tests/claude-session.test.ts imports it by name.
- Fifteen comments cited paths under `docs/`, deleted in #1083. Each keeps its
  explanation and drops the dead pointer.
